automation/python/requirements.txt
okta-sdk-python==2.9.5
python-dotenv==1.0.0
requests==2.31.0
